Rewrite core parser logic to fix stack overflow
Using `cargo fuzz`, I figured out that it was possible to cause a stack
overflow with the old recursive approach. I took this as an opportunity
to split the lexing from the parsing to generate a cleaner solution.

This commit starts work on issue #7. By making the parser itself
non-recursive, stack overflows now don't happen during parsing.

This doesn't fix issue #7 though since `SgfNodeBuilder::build` is still
recursive, and at the end of parsing, that can lead to a stack overflow.