Rewrite SgfNodBuilder::build to avoid recursion
The simple recursive algorithm could result in a stack overflow if there
were enough children. This approach is more complex, and less efficient
with the heap allocations for the `Rc`s, but it's algorithmically
efficient, should be fast enough for any real world uses, and should
never result in a stack overflow.

Fixes #7.