Finish and test sgf validation
This finishes the work on validation. I'll eventually want to rewrite
the validation logic to be non-recursive, but that change will be a
non-breaking change that can be made for 3.0.1 or something similar.

At this point, the only blockers to 3.0 release are handling game
specific properties, and testing the library in a real project.