Clean up docs
This commit standardizes the use of links in documentation, adds module
level docs for the `go` and `unknown_game` modules, and cleans up the
`go::Prop` and `unknown_game::Prop` documentation to be a little
clearer.

I've also removed a stray `pub` on `general_validate_properties` which
was never intended to be public! Technically this is a breaking change,
but I suspect no one will mind if I quietly remove it.