Fix gametype error handling
The old logic treated "1" as Go (which was the default), and accepted
any other string as Unknown.

This is subtly wrong though. If the value can't be parsed as a number,
then it should return an `Invalid` property.

The other change here is that in lenient mode, if we can't parse the
game type, we'll assume it's Go.